# Convert f32 binary expression tests to use pre-generated inputs