--- a/src/webgpu/shader/execution/expression/binary/f32_arithmetic.spec.ts
+++ b/src/webgpu/shader/execution/expression/binary/f32_arithmetic.spec.ts
@@ -12,14 +12,12 @@ import {
   remainderInterval,
   subtractionInterval,
 } from '../../../../util/f32_interval.js';
-import { cartesianProduct, fullF32Range } from '../../../../util/math.js';
+import { kVectorTestValues } from '../../../../util/math.js';
 import { allInputSources, Case, makeBinaryToF32IntervalCase, run } from '../expression.js';
 
 import { binary } from './binary.js';
 
 export const g = makeTestGroup(GPUTest);
-
-const kTestValues: number[][] = cartesianProduct(fullF32Range(), fullF32Range());
 
 g.test('addition')
   .specURL('https://www.w3.org/TR/WGSL/#floating-point-evaluation')
@@ -37,7 +35,7 @@ Accuracy: Correctly rounded
       return makeBinaryToF32IntervalCase(lhs, rhs, additionInterval);
     };
 
-    const cases = kTestValues.map((v: number[]) => {
+    const cases = kVectorTestValues[2].map(v => {
       return makeCase(v[0], v[1]);
     });
 
@@ -60,7 +58,7 @@ Accuracy: Correctly rounded
       return makeBinaryToF32IntervalCase(lhs, rhs, subtractionInterval);
     };
 
-    const cases = kTestValues.map((v: number[]) => {
+    const cases = kVectorTestValues[2].map(v => {
       return makeCase(v[0], v[1]);
     });
 
@@ -83,7 +81,7 @@ Accuracy: Correctly rounded
       return makeBinaryToF32IntervalCase(lhs, rhs, multiplicationInterval);
     };
 
-    const cases = kTestValues.map((v: number[]) => {
+    const cases = kVectorTestValues[2].map(v => {
       return makeCase(v[0], v[1]);
     });
 
@@ -106,7 +104,7 @@ Accuracy: 2.5 ULP for |y| in the range [2^-126, 2^126]
       return makeBinaryToF32IntervalCase(lhs, rhs, divisionInterval);
     };
 
-    const cases = kTestValues.map((v: number[]) => {
+    const cases = kVectorTestValues[2].map(v => {
       return makeCase(v[0], v[1]);
     });
 
@@ -129,7 +127,7 @@ Accuracy: Derived from x - y * trunc(x/y)
       return makeBinaryToF32IntervalCase(lhs, rhs, remainderInterval);
     };
 
-    const cases = kTestValues.map((v: number[]) => {
+    const cases = kVectorTestValues[2].map(v => {
       return makeCase(v[0], v[1]);
     });